fix: remove rootmodal to new folder
