+++ cypress/support/commands/item.ts
@@ -14,6 +14,7 @@ import {
   SHARE_ITEM_EMAIL_INPUT_ID,
   SHARE_ITEM_SHARE_BUTTON_ID,
   TREE_MODAL_CONFIRM_BUTTON_ID,
+  buildHomeModalItemID,
   buildItemFormAppOptionId,
   buildItemRowArrowId,
   buildPermissionOptionId,
@@ -57,7 +58,7 @@ Cypress.Commands.add(
         // click on the element
         if (idx === array.length - 1) {
           cy.wrap($tree)
-            .get(`#${buildTreeItemId(value, treeRootId)}`)
+            .get(`#${buildHomeModalItemID(value)}`)
             .first()
             .click();
         }
@@ -67,7 +68,7 @@ Cypress.Commands.add(
           !$tree.find(`#${buildTreeItemId(array[idx + 1], treeRootId)}`).length
         ) {
           cy.wrap($tree)
-            .get(`#${buildTreeItemId(value, treeRootId)}`)
+            .get(`#${buildHomeModalItemID(value)}`)
             .trigger('mouseover')
             .get(`#${buildItemRowArrowId(value)}`)
             .first()
