+++ src/langs/en.json
@@ -291,7 +291,7 @@
   "SHORT_LINK_MAX_CHARS_ERROR": "The short link must not exceed {{data}} chars.",
   "SHORT_LINK_INVALID_CHARS_ERROR": "The short link contains invalid chars ({{data}}).",
   "SHORT_LINK_UNKNOWN_ERROR": "An unknown error occurred.",
-  "EMPTY_FOLDER_CHILDREN_FOR_THIS_ITEM": "No Sub-Folders For this Item",
+  "EMPTY_FOLDER_CHILDREN_FOR_THIS_ITEM": "This item does not contain any folder.",
   "TO": "to",
   "ROOT": "Root"
 }
